Fix broken imports after moving files
